requirements_simple.txt
# Essential dependencies for CNN Fear & Greed Index Bot

# Telegram Bot Framework
python-telegram-bot>=20.0,<21.0

# HTTP Requests
aiohttp>=3.8.0
requests>=2.28.0

# Database
SQLAlchemy>=2.0.0
aiosqlite>=0.19.0

# Environment Variables
python-dotenv>=1.0.0

# Web Scraping (for backup data sources)
beautifulsoup4>=4.12.0

# Date/Time handling
python-dateutil>=2.8.0 